+++ django_energysystem_viewer/urls.py
@@ -6,19 +6,18 @@ app_name = "django_energysystem_viewer"
 
 
 urlpatterns = [
-    path("energysystem/network/", views.network),
+    path("energysystem/selection/", views.SelectionView.as_view(), name="selection"),
+    path("energysystem/network/", views.network, name="networks"),
     path("energysystem/network_graph/", views.network_graph),
-    path("energysystem/abbreviation/", views.abbreviations),
+    path("energysystem/abbreviation/", views.abbreviations, name="abbreviations"),
     path("energysystem/abbreviation_meaning/", views.abbreviation_meaning),
-    path("energysystem/aggregation/", views.AggregationView.as_view()),
-    path("energysystem/aggregation_graph/", views.aggregation_graph),
-    path("collections/", views.CollectionsView.as_view(), name="collections"),
-    path("collection/<str:collection_name>/processes/", views.ProcessesView.as_view(), name="processes"),
+    path("energysystem/aggregation/", views.AggregationView.as_view(), name="aggregations"),
+    path("collection/processes/", views.ProcessesView.as_view(), name="processes"),
     path(
         "collection/<str:collection_name>/process/<str:process_name>/data/",
         views.ProcessDetailView.as_view(),
     ),
-    path("collection/<str:collection_name>/artifacts/", views.ArtifactsView.as_view(), name="artifacts"),
+    path("collection/artifacts/", views.ArtifactsView.as_view(), name="artifacts"),
     path(
         "collection/<str:collection_name>/artifact/<str:group_name>/<str:artifact_name>/data/",
         views.ArtifactDetailView.as_view(),